Add files via upload
The library in a village called Mountain Caves need a new electronic rental system and it is up to you to build it. Mountain Caves has two libraries. Each library offers many books to rent. Customers can print the list of available books, borrow, and return books. There are two classes Book and Library, that provide the functionality for the book database. You must implement methods to make the classes work. Book title, find out if the book is available, borrow the book, and return the book. On the library class we are dealing with managing collection of books, All the Libraries have the same hours 9AM to 5PM daily. However, they vary in addresses and book collections
As the owner of a Gardens hotel management system, you've tasked your software development team with creating a program to facilitate room reservations. The program should be able to handle multiple hotels, each with various rooms of different capacities and prices. Customers will be able to make reservations for specific dates. Design the classes and methods for this program, considering the essential functionalities required for hotel room booking.